% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_LTFH.R
\name{Assign_LTFH}
\alias{Assign_LTFH}
\title{Assign_LTFH function}
\usage{
Assign_LTFH(Pheno_data, valT, h2, with_sib = 1, Burn_in = 100)
}
\arguments{
\item{Pheno_data}{Is the Phenotype data that comes out of the simulation when running SimulerData}

\item{valT}{is the dignificance level, usually 0.05}

\item{h2}{Is the .. usually 0.5}

\item{with_sib}{Binary, defines wheter or not we should include siblings. Defaults to 1 = yes.}

\item{Burn_in}{Defines the number of samples to discard in the gibs-sampling. Defaults to 100 (found by simulating convergence tests)}
}
\description{
This function assigns phenotypes based on LTFH article. This implementation uses Gibs-sampling to efficiently assign genetic liabilities to each individual.
}
\examples{
Assing_LTFH((Pheno_data, valT, h2, with_sib = 1, Burn_in = 100)
}
\keyword{Assign}
\keyword{LTFH}
