% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_hat_sigma_hat}
\alias{beta_hat_sigma_hat}
\title{MLE and standard deviation of SNP effect size in genetic LMM}
\usage{
beta_hat_sigma_hat(K, X, y, log_lambda_tau_hat)
}
\arguments{
\item{K}{The kinship matrix in genetic LMM}

\item{X}{The augmented matrix of fixed effects which is obtained by
augmenting the genotype vector of the SNP to the covariate matrix}

\item{y}{The phenotype vector (should be a column matrix)}

\item{log_lambda_tau_hat}{The \eqn{2}-vector consisting of the log of 
RMLE estimates of \eqn{\lambda} and \eqn{\tau}}
}
\value{
A vector of two elements where the first element is the MLE,
and the second element is the standard deviation of the SNP effect size
}
\description{
\eqn{beta_hat_sigma_hat} returns the maximum likelihood estimate of 
SNP effect size and its standard deviation.
}
\details{
The MLE and standard deviation of SNP effect size is used in performing 
the hypothesis testing of no association. For more details, see our article.
}
