% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareBox.R
\name{CompareBox}
\alias{CompareBox}
\title{Boxplot for comparisons}
\usage{
CompareBox(sib)
}
\arguments{
\item{sib}{Binary indicator to indicate if we want to look at siblings}
}
\description{
This function plots 3 boxplots to compare the 3 methods
}
\examples{
CompareBox(sib = 1)
}
\keyword{Boxplots}
