% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDistribution.R
\name{ConditionalDistribution}
\alias{ConditionalDistribution}
\title{ConditionalDistribution function}
\usage{
ConditionalDistribution(CovMatrix)
}
\arguments{
\item{CovMatrix}{Is the covariance matrix of the multivariate normal distribution}
}
\description{
This function calculates the mean constant and the variance of the normal conditional distributions. Output is a list of the two vector/matrix
}
\examples{
ConditionalDistribution(matrix(ncol = 4, nrow = 4))
}
\keyword{Conditional}
\keyword{Distributions}
