% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrixLtfh.R
\name{ConfusionMatrixLtfh}
\alias{ConfusionMatrixLtfh}
\title{ConfusionMatrixLtfh}
\usage{
ConfusionMatrixLtfh(total_indiv, SNP, h, sib, BFC)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{sib}{Is a binary for indicating if we look at siblings}

\item{BFC}{Is a variable defining the size of the devider (BFC)}
}
\description{
This function plots confusion matrix for the LT-FH method
}
\examples{
ConfusionMatrixLtfh(total_indiv = 1000, SNP = 1000, h = 0.5, sib =1, BFC = 1)
}
\keyword{ConfusionMatrix}
