% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrixTG.R
\name{ConfusionMatrixTG}
\alias{ConfusionMatrixTG}
\title{ConfusionMatrixTG}
\usage{
ConfusionMatrixTG(total_indiv, SNP, h, BFC)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{BFC}{Is a variable defining the size of the devider (BFC)}
}
\description{
This function plots confusion matrix for the True Genetic Liability
}
\examples{
ConfusionMatrixTG(total_indiv = 1000, SNP = 1000, h = 0.5, BFC = 1)
}
\keyword{ConfusionMatrix}
