% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBoxData.R
\name{CreateBoxData}
\alias{CreateBoxData}
\title{Creating data for boxplot}
\usage{
CreateBoxData(sib)
}
\arguments{
\item{sib}{Binary indicator to indicate if we want to look at siblings}
}
\description{
This function creates data used to plot 3 boxplots to compare the 3 methods
}
\examples{
CreateBoxData(sib = 1)
}
\keyword{BoxplotsData}
