% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTransMatrix.R
\name{CreateTransMatrix}
\alias{CreateTransMatrix}
\title{CreateTransMatrix function}
\usage{
CreateTransMatrix(index, covM)
}
\arguments{
\item{index}{Defines the index in the vector we will bring to the top}

\item{covM}{Defines the CovarianceMatrix for the distribution}
}
\description{
This function creates a transformation matrix used to calculate conditional distributions
}
\examples{
CreateTransMatrix(1,matrix(ncol = 4, nrow = 4))
}
\keyword{transformation}
