% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_tilda}
\alias{gamma_tilda}
\title{MLE of fixed effect sizes under known variance components}
\usage{
gamma_tilda(K, X, y, log_lambda_tau)
}
\arguments{
\item{K}{The kinship matrix in genetic LMM}

\item{X}{The augmented matrix of fixed effects which is obtained by
augmenting the genotype vector of the SNP to the covariate matrix}

\item{y}{The phenotype vector (should be a column matrix)}

\item{log_lambda_tau}{The vector \eqn{(log \lambda, log \tau) }, where \eqn{\lambda}
and \eqn{\tau} are the two parameters used in the variance components 
of the genetic LMM.}
}
\value{
MLE of \eqn{\gamma}, the vector of fixed effect sizes 
for known value of \eqn{(\lambda, \tau)}
}
\description{
\eqn{gamma_tilda} returns the MLE of covariate and SNP effect sizes
in a genetic LMM when the variance components of polygenic
effect and observational noise are known.
}
\details{
The MLE of covariate and SNP effect sizes for known values of \eqn{\lambda}
and \eqn{\tau} is used in finding the RMLE of \eqn{(\lambda, \tau)}
by maximizing the marginal likelihood. This in turn helps to find the unrestricted MLE
of SNP effect size. For more details, see our article.
}
