% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logreg.R
\name{get_lr_stats}
\alias{get_lr_stats}
\title{Calculate results from logistic regression}
\usage{
get_lr_stats(x, y, df, n = 0)
}
\arguments{
\item{x}{binary predictor data, e.g., for gene mutation or SNP}

\item{y}{binary response data}

\item{df}{\code{data.frame} containing \code{x} and \code{y}}
}
\value{
a single row \code{data.frame} containing slope, the standard dev., and p-value.
Suitable for usage with \code{\link[=msg_catch_lapply]{msg_catch_lapply()}}.
}
\description{
This function returns the slope, standard deviation, and p-value
from logistic regression calculated using \code{glm()}.
}
\details{
Logistic regression function: \code{glm(y ~ x, family = "binomial", data = df)}

The formula for logistic regression uses
}
