% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oddsratio.R
\name{get_oddsratio}
\alias{get_oddsratio}
\title{Get odds ratio statistics}
\usage{
get_oddsratio(x, y)
}
\arguments{
\item{x}{the predictor vector, binary}

\item{y}{the outcome vector, binary}
}
\value{
Single row data frame that can be used with \code{\link[=msg_catch_lapply]{msg_catch_lapply()}}
}
\description{
Given a gene, wraps \code{epitools::oddsratio} to return the odds ratio
confidence interval, the Fisher exact result, the Chi-squared results,
and the p-value estimation.
}
