% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanSD.R
\name{getMeanSD}
\alias{getMeanSD}
\title{Get Mean and SD function}
\usage{
getMeanSD(total_indiv, SNP, h, Child, Mom, Dad, Nr_sib = 0, Sib_status = 0)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{Child}{Is a binary for indicating case/notcase for child}

\item{Mom}{Is a binary for indicating case/notcase for Mom}

\item{Dad}{Is a binary for indicating case/notcase for Dad}

\item{Sib_status}{indicates the number of sib-cases}

\item{Nr-Sib}{indiacates the number of siblings for the family}
}
\description{
This function grabs the mean and sd from the genetic-liability sampled using the Assign_LTFh fuction
}
\examples{
getMeanSD(total_indiv, SNP, h, Child, Mom, Dad, Nr_sib=0, Sib_status=0)
}
\keyword{Mean/SD}
