% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KillSib.R
\name{KillSib}
\alias{KillSib}
\title{KillSib function}
\usage{
KillSib(true)
}
\arguments{
\item{true}{Is a TRUE-file from simulation.}
}
\description{
This function assigns the number of siblings to each family-id based on a maksimum likelihood estimate for data on DST.dk
}
\examples{
KillSib(true = tibble::tibble(ID = 1:100, Child = rbinom(100,1,0.3), Mom = rbinom(100,1,0.3), Dad = rbinom(100,1,0.3), Nr_sib = rep(0,100), Sib_status = rep(0,100), Sib1 = rep(0,100), Sib2 = rep(0,100), Sib3 = rep(0,100), Sib4 = rep(0,100), Sib5 = rep(0,100), Sib6 = rep(0,100), Sib7 = rep(0,100)))
}
\keyword{Siblings}
