% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregressionplotCC.R
\name{linearregressionplotcc}
\alias{linearregressionplotcc}
\title{LinearRegressionPlot (Beta)}
\usage{
linearregressionplotcc(SNPno, total_indiv, SNP, h, sib)
}
\arguments{
\item{SNPno}{The specific SNP we are looking at}

\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{sib}{Is a binary for indicating if we look at sibling history}
}
\description{
This function plots linearregresion plot to illustrate beta for case_ctrl
}
\examples{
linearregressionplotcc(SNPno = 10, total_indiv = 1000, SNP = 1000, h = 0.5, sib=0)
}
\keyword{regplot}
