% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregressionplotltfh.R
\name{linearregressionplotltfh}
\alias{linearregressionplotltfh}
\title{LinearRegressionPlot (Beta)}
\usage{
linearregressionplotltfh(SNPno, total_indiv, SNP, h, sib)
}
\arguments{
\item{SNPno}{The specific SNP we are looking at}

\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{sib}{Is a binary for indicating if we look at sibling history}
}
\description{
This function plots linearregresion plot to illustrate beta
}
\examples{
linearregressionplotltfh(SNPno = 10, total_indiv = 1000, SNP = 1000, h = 0.5, Child=1, Mom=1, Dad=1, Nr_sib=0, Sib_status=0)
}
\keyword{DistPlot}
