% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTransformation.R
\name{LinearTransformation}
\alias{LinearTransformation}
\title{LinearTransformation function}
\usage{
LinearTransformation(covM)
}
\arguments{
\item{covM}{Is a covariance matrix}
}
\description{
This function makes a linear transformation of a CovarianceMatrix.
}
\examples{
LinearTransformation(covM)
}
\keyword{Linaer}
\keyword{Transformation}
