% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_catch.R
\name{log_catch}
\alias{log_catch}
\alias{log_catch_lapply}
\title{Factory for logged tryCatch}
\usage{
log_catch(fun)

log_catch_lapply(fun, X)
}
\arguments{
\item{fun}{function to convert}

\item{...}{additional arguments to pass to the function}
}
\description{
Converts a function into a tryCatch version that logs the result.
Taken from Martin Morgan's answer to https://stackoverflow.com/questions/4948361/.
\code{log_catch_lapply} is identical except the tryCatch also applies a \code{lapply}.
}
\section{Functions}{
\itemize{
\item \code{log_catch_lapply}: An application of \code{log_catch} to lapply.
}}

