% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pc_formula.R
\name{make_pc_formula}
\alias{make_pc_formula}
\title{Make a formula with a specified number of principal components}
\usage{
make_pc_formula(x, y = "culture_2m", n = 0)
}
\arguments{
\item{x}{the name of the predictor}

\item{y}{the name of the response}

\item{n}{the number of principal components.}
}
\description{
This function is a helper for \code{\link[=get_logreg]{get_logreg()}}.
}
