% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_temp_labels.R
\name{make_temp_labels}
\alias{make_temp_labels}
\title{Helper to label vector elements}
\usage{
make_temp_labels(x)
}
\arguments{
\item{x}{a vector}
}
\value{
\code{x} with names that are the elements of \code{x}
}
\description{
Helper for \code{\link[=msg_catch_lapply]{msg_catch_lapply()}}. Labels are
}
