% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan.R
\name{manhattan.circular}
\alias{manhattan.circular}
\title{Circular Manhattan plot                                                     
                                                                             
Some text here}
\usage{
manhattan.circular(pVal, pos, chr, annot, log = FALSE, filename = NULL,
  width = 10, height = 10)
}
\arguments{
\item{pVal}{Vector of p-values}

\item{pos}{Vector of genomic positions for each p-value}

\item{chr}{Corresponding vector of chromosomes for p-value}

\item{annot}{annotation for some of the top hits (not supported yet)}

\item{log}{True if p-values are on logarithmic form. False as default.}

\item{filename}{Name if saving to file}

\item{width}{Width when saving plot}

\item{height}{Height when saving plot}
}
\value{
returns a ggplot2 object, piped to X11
}
\description{
Circular Manhattan plot                                                     
                                                                             
Some text here
}
