% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_catch_lapply.R
\name{msg_catch_lapply}
\alias{msg_catch_lapply}
\title{Message catching over a list}
\usage{
msg_catch_lapply(fun, X, logfile = NULL, ...)
}
\arguments{
\item{fun}{the function to be passed to \code{msg_catch}}

\item{X}{the argument to \code{lapply} over}

\item{logfile}{leave \code{NULL} to not log. Otherwise, a character string/connection to the log file.}

\item{...}{additional arguments to be passed to \code{fun}}
}
\value{
list \code{fun_list(res, warn, err)} that contains a \code{data.frame} of results,
a character vector of warnings, and a chr vector of errors.
}
\description{
Integration of \code{msg_catch} and \code{lapply}. Basically a wrapper for \code{lapply}
paplied to a \code{msg_catch} function, followed by a \code{msg_catch_relist} and
optional \code{write_log}.
}
\details{
Make sure functions passed to \code{fun} output a \code{data.frame} that can be passed
to \code{rbind}.
}
