% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_catch_relist.R
\name{msg_catch_relist}
\alias{msg_catch_relist}
\title{Helper functions for log_catch lists}
\usage{
msg_catch_relist(x)
}
\arguments{
\item{x}{the output from \code{lapply()} with \code{\link[=msg_catch]{msg_catch()}}. Is a list of lists of the form \code{list(res, warn = w, err = e)}}
}
\description{
After a \code{\link[=msg_catch]{msg_catch()}} product function has been used with \code{lapply()},
the results, warning, and errors can be resorted into separate lists.
}
