% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neglr_lambda_tau}
\alias{neglr_lambda_tau}
\title{Marginal likelihood of variance parameters in genetic LMM}
\usage{
neglr_lambda_tau(log_lambda_tau, K, X, y)
}
\arguments{
\item{log_lambda_tau}{The vector \eqn{(log \lambda, log \tau) }, where \eqn{\lambda}
and \eqn{\tau} are the two parameters used in the variance components 
of the genetic LMM.}

\item{K}{The kinship matrix in genetic LMM}

\item{X}{The augmented matrix of fixed effects which is obtained by
augmenting the genotype vector of the SNP to the covariate matrix}

\item{y}{The phenotype vector (should be a column matrix)}
}
\value{
Marginal likelihood of \eqn{(\lambda, \tau)}
}
\description{
\eqn{neglr_lambda_tau} returns the (negated) marginal likelihood
of the parameters used in the variance component of the 
genetic LMM
}
\details{
The marginal likelihood obtained is minimized (since it is negated)
to find the RMLE of \eqn{(\lambda, \tau)}.
For more details, see our article.
}
