% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_data.R
\name{Normalize_data}
\alias{Normalize_data}
\title{Assing GWAX function}
\usage{
Normalize_data(df, MAF)
}
\arguments{
\item{df}{Is a matrix/data-frame that we want to Normalize (number_of_indivs X SNPs)}

\item{MAF}{Defines the Minor Allele Frequancy, and is a vector of probabilities from unif(0.01,0.49) one for each SNP}
}
\description{
This function assigns phenotypes based on GWAX.
}
\examples{
Normalize_data()
}
\keyword{Normalize}
