% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_stats.R
\name{or_stats}
\alias{or_stats}
\title{Get odds ratio statistics}
\usage{
or_stats(x, y)
}
\arguments{
\item{x}{the predictor vector, binary}

\item{y}{the outcome vector, binary}
}
\value{
Single row data frame (used for row binding)
}
\description{
Given a gene, wraps \code{epitools::oddsratio} to return the odds ratio
confidence interval, the Fisher exact result, the Chi-squared results,
and the p-value estimation.
}
