% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plosteriot.R
\name{plosteriot}
\alias{plosteriot}
\title{Plot distributions}
\usage{
plosteriot(total_indiv, SNP, h, Child, Mom, Dad, Nr_sib = 0, Sib_status = 0)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{Child}{Is a binary for indicating case/notcase for child}

\item{Mom}{Is a binary for indicating case/notcase for Mom}

\item{Dad}{Is a binary for indicating case/notcase for Dad}

\item{Sib_status}{indicates the number of sib-cases defaults to 0}

\item{Nr-Sib}{indiacates the number of siblings for the family defaults to 0}
}
\description{
This function plots posterior distribution and prior
}
\examples{
plosteriot(total_indiv, SNP, h, Child, Mom, Dad, Nr_sib=0, Sib_status=0)
}
\keyword{DistPlot}
