% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{QQ}
\alias{QQ}
\title{QQ plot}
\usage{
QQ(x, main = "", file = NULL, ncol = 2, width = NULL,
  height = NULL, overlay = FALSE, free = "free")
}
\arguments{
\item{x}{A result object from the EWAS, of class "MoBa.epigenetic.result"}

\item{file}{Filename if QQ plot is to be save as a FPD/JPEG. NOTE file extention 
should not be included in the filename.}

\item{ncol}{Number of colunms if the QQ plot spans a matrix layout}

\item{width}{Width when saving plot}

\item{height}{Height when saving plot}

\item{overlay}{If multiple QQ plots is to be overlayed instead of ploted in seperate windows}
}
\value{
returns a ggplot2 object, piped to X11
}
\description{
Print out the top CpGs from an EWAS model, with three collumns,
the effect size, the standard deviation and the p-value for each
CpG. The CpGs are sorted by p-value if sort = TRUE.
}
