% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relist_helper.R
\name{relist_helper}
\alias{relist_helper}
\title{Helper for relisting}
\usage{
relist_helper(x, x_subset, msg_label = NULL)
}
\arguments{
\item{x}{an output from using \code{lapply()} and a \code{\link[=msg_catch]{msg_catch()}} function}

\item{x_subset}{the name of the subset to make into a new list}

\item{msg_label}{an optional character string to start each entry}
}
\description{
A helper function that remakes a list based on a given subset name. Only used for message (warnings and errors).
}
