% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Standard Error of the Mean}
\usage{
sem(x)
}
\arguments{
\item{x}{vector of numbers}
}
\description{
This function calculates the standard error of a vector of means
}
\examples{
sem(x)
}
\keyword{Error}
\keyword{Standard}
