% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulerData.R
\name{SimulerData}
\alias{SimulerData}
\title{SimulerData function}
\usage{
SimulerData(total_indiv, indiv_chunk, SNP, h, c, k, nr_workers)
}
\arguments{
\item{total_indiv}{The total number of individuals to simulate}

\item{indiv_chunk}{The number of individuals to simulate in each iteration (should not be above 5000)}

\item{SNP}{The number of SNPs to simulate}

\item{h}{the h^2 - meaning the .. ususally 0.5}

\item{c}{The number of causal SNPs in the simulaiton usually 1/10000*SNP}

\item{k}{Significance level - usually 0.05}

\item{nr_workers}{Number of cores to run the code in parallel}
}
\description{
This function simulates genetic-data and saves relevant files doing so.
}
\examples{
SimulerData(1000,100,1000,0.5,100,0.05,2)
}
\keyword{Genetic}
\keyword{simulation}
