% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformToPlink.R
\name{TransformToPlink}
\alias{TransformToPlink}
\title{Transform to PLINK function}
\usage{
TransformToPlink(df, indiv_chunk)
}
\arguments{
\item{df}{Is a matrix/data-frame that we want to Normalize (number_of_indivs X SNPs)}

\item{indiv_chunk}{Defines the number of individuals in the dataframe}
}
\description{
This function takes a matrix of 0,1,2 (genotypes) and transforms it to data readable to PLINK-software.
}
\examples{
TransformToPlink(df = matrix(ncol = 10, nrow = 10), indiv_chunk = 10)
}
\keyword{PLINK}
