% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{V_lambda_tau}
\alias{V_lambda_tau}
\title{Variance-covariance matrix of random components in LMM}
\usage{
V_lambda_tau(K, log_lambda_tau)
}
\arguments{
\item{K}{The kinship matrix in genetic LMM}

\item{log_lambda_tau}{The vector \eqn{(log \lambda, log \tau) }, where \eqn{\lambda}
and \eqn{\tau} are the two parameters used in the variance components 
of the genetic LMM.}
}
\value{
\eqn{(1/\tau)(\lambda K + I)    }
}
\description{
\eqn{V_lambda_tau} returns the variance-covariance matrix
of the sum of polygenic effects and observational noise
in a genetic LMM
}
\details{
The variance covariance matrix is used in finding the 
RMLE of \eqn{\lambda} and \eqn{\tau}. This eventually 
helps us to find the MLE of SNP effect size. For more details,
see our article.
}
