% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{Write a log file from caught messages}
\usage{
write_log(x, f)
}
\arguments{
\item{x}{the output from a \code{msg_catch_relist} call. Requires errors to be a character vector under \code{.$err} and warnings to be a vector under \code{.$warn}.}

\item{f}{character string or file location of the log file.}
}
\description{
After messages have been caught by a \code{msg_catch} function, this function will
write warnings and errors into a text file.
}
