% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assign_GWAX.R
\name{Assign_GWAX}
\alias{Assign_GWAX}
\title{Assing GWAX function}
\usage{
Assign_GWAX(true, with_sib = 1)
}
\arguments{
\item{true}{Is a TRUE-file from simulation.}

\item{with_sib}{Defines if siblings Default is 1=true, else 0}
}
\description{
This function assigns phenotypes based on GWAX.
}
\examples{
Assign_GWAX(tibble::tibble(ID = 1:100, Child = rbinom(100,1,0.3), Mom = rbinom(100,1,0.3), Dad = rbinom(100,1,0.3), Nr_sib = rep(0,100), Sib_status = rep(0,100)), with_sib = 1)
}
\keyword{gwax}
