% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceMatrix.R
\name{CovarianceMatrix}
\alias{CovarianceMatrix}
\title{covariance matrix function}
\usage{
CovarianceMatrix(h2, nr_sib = 0)
}
\arguments{
\item{nr_sib}{Defines the number of sibling, defaults to 0}

\item{h^2}{is the .. usually 0.5}
}
\description{
This function calculates the covariance matrix based on h^2 and the number of siblings
}
\examples{
CovarianceMatrix()
}
\keyword{covariance}
\keyword{matrix}
