% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanSD.R
\name{getMeanSD}
\alias{getMeanSD}
\title{Get Mean and SD function}
\usage{
getMeanSD(
  total_indiv,
  SNP,
  h,
  Child_i,
  Mom_i,
  Dad_i,
  Nr_sib_i = 0,
  Sib_status_i = 0
)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{Child_i}{Is a binary for indicating case/notcase for child}

\item{Mom_i}{Is a binary for indicating case/notcase for Mom}

\item{Dad_i}{Is a binary for indicating case/notcase for Dad}

\item{Nr_sib_i}{indiacates the number of siblings for the family}

\item{Sib_status_i}{indicates the number of sib-cases}
}
\description{
This function grabs the mean and sd from the genetic-liability sampled using the Assign_LTFh fuction
}
\examples{
getMeanSD(total_indiv = 1000, SNP = 1000, h = 0.5, Child = 1, Mom = 1, Dad = 1, Nr_sib=0, Sib_status=0)
}
\keyword{Mean/SD}
