% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KillSib.R
\name{KillSib}
\alias{KillSib}
\title{KillSib function}
\usage{
KillSib(true)
}
\arguments{
\item{true}{Is a TRUE-file from simulation.}
}
\description{
This function assigns the number of siblings to each family-id based on a maksimum likelihood estimate for data on DST.dk
}
\examples{
KillSib(true)
}
\keyword{Siblings}
