% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanBigltfh.R
\name{manhattanBigltfh}
\alias{manhattanBigltfh}
\title{Manhattan Big Plot LT-FH}
\usage{
manhattanBigltfh(total_indiv, SNP, h, sib = 0)
}
\arguments{
\item{total_indiv}{Number of indicviduls in the dataset}

\item{SNP}{Is the number of SNPs in the data}

\item{h}{Is the heritability usually 0.5}

\item{sib}{Is a binary for indicating if we look at sibling history}
}
\description{
This function plots a manhatten plot with tooltip
}
\examples{
manhattanBigCC(total_indiv = 1000, SNP = 1000, h = 0.5, sib=0)
}
\keyword{Manhattan}
