% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_catch.R
\name{msg_catch}
\alias{msg_catch}
\title{Factory for tryCatch with message retention}
\usage{
msg_catch(fun)
}
\arguments{
\item{fun}{function to convert}

\item{...}{additional arguments to pass to the function}
}
\description{
Converts a function into a tryCatch version that records status messages.
Taken from Martin Morgan's answer to https://stackoverflow.com/questions/4948361/.
\code{\link[=msg_catch_lapply]{msg_catch_lapply()}} wraps this with a \code{lapply()} and \code{\link[=msg_catch_relist]{msg_catch_relist()}}.
}
