% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_catch_lapply.R
\name{msg_catch_lapply}
\alias{msg_catch_lapply}
\title{Message catching over a list}
\usage{
msg_catch_lapply(fun, X, logfile = NULL, ...)
}
\arguments{
\item{fun}{the function to be passed to \code{\link[=msg_catch]{msg_catch()}}}

\item{X}{the argument to \code{lapply()} over}

\item{logfile}{leave \code{NULL} to not log. Otherwise, a character string/connection to the log file.}

\item{...}{additional arguments to be passed to \code{fun}}
}
\value{
list \code{fun_list(res, warn, err)} that contains a \code{data.frame} of results,
a character vector of warnings, and a chr vector of errors.
}
\description{
Integration of \code{\link[=msg_catch]{msg_catch()}} and \code{lapply()}. Wraps together a \code{lapply()} that
takes a \code{\link[=msg_catch]{msg_catch()}} function and a \code{\link[=msg_catch_relist]{msg_catch_relist()}}.
Optional \code{logfile} argument can call \code{\link[=write_log]{write_log()}}.
}
\details{
Make sure functions passed to \code{fun} have an output that can be passed to \code{rbind()},
such as a single row a \code{data.frame}.
}
